+++ config/test.exs
@@ -1,0 +1,3 @@
+import Config
+
+config :aws_credentials, credential_providers: []